compat with diffusers 1.14

+++ main.py
@@ -221,7 +221,7 @@ def randomSentenceV2 ():
 
 def main (argv):
 	global pipe, senGen2, senGen
-	pipe = StableDiffusionPipeline.from_pretrained(DIFFUSER_MODEL_PATH, use_auth_token=HF_TOKEN)
+	pipe = StableDiffusionPipeline.from_pretrained(DIFFUSER_MODEL_PATH, use_auth_token=HF_TOKEN, torch_dtype=torch.float32)
 
 	device = torch.device(f'{DEVICE}:{TEXT_DEVICE_INDEX}') if DEVICE else None
 	senGen = SentenceGenerator(templates_path='corpus/templates.txt', reserved_path='corpus/reserved.txt', device=device)
